refactor: extract AgentFormModal from AgentList.tsx
Move the agent form (fields, tool toggles, skill management) into a
dedicated component. AgentList.tsx reduced from 891 to 424 lines.
